refactor: move LibTester
